Refactor sgl kernel folder organization
