+++ sgl-kernel/csrc/allreduce/custom_all_reduce_hip.cuh
