+++ sgl-kernel/csrc/speculative/eagle_utils.cu
