+++ sgl-kernel/sgl_kernel/attention.py
@@ -1,4 +1,3 @@
-import sgl_kernel.ops._kernels
 import torch
